Add IGuildChannelContainer and getChannelById(Class, id)
