+++ src/main/java/net/dv8tion/jda/api/entities/ChannelType.java
@@ -144,7 +144,7 @@ public enum ChannelType
             case GROUP:
                 return true;
             default:
-                return false;
+                return isThread();
         }
     }
 
@@ -154,7 +154,8 @@ public enum ChannelType
      *
      * @return True, if channels of this type are {@link ThreadChannel ThreadChannel}
      */
-    public boolean isThread() {
+    public boolean isThread()
+    {
         switch (this)
         {
             case GUILD_NEWS_THREAD:
